# docs/openapi.v1.json
{
  "components": {
...
    "securitySchemes": {
      "mobileSessionAuth": {
        "description": "Mobile-only opaque session transport. Exact header form: Authorization: Session <opaque-token>. Not a Bearer token and not accepted for web cookie sessions.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "recoveryGrantAuth": {
        "description": "Restricted recovery-grant authorization for recovery passkey registration. Exact header form: Authorization: RecoveryGrant <opaque-token>. Not an account session and not a Bearer token.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "sessionAuth": {
        "description": "Web-only opaque session cookie. The runtime sets it as Secure, HttpOnly, SameSite=Lax, Path=/, with no Domain. Mutative cookie-authenticated routes require same-origin/same-site CSRF evidence.",
        "in": "cookie",
        "name": "__Host-Http-town_session",
        "type": "apiKey"
      },
      "setupGrantAuth": {
        "description": "Restricted setup-grant authorization for first-passkey registration. Exact header form: Authorization: SetupGrant <opaque-token>. Not an account session and not a Bearer token.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "TownControlKey": {
        "description": "Temporary controlled test mechanism for confirmation routes. Not public authentication. Do not treat this as a user session, OAuth token, or production identity system. Example secret values are intentionally omitted.",
        "in": "header",
        "name": "X-TOWN-Control-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "TOWN API civic foundation — health probes, communities, published signals, temporary controlled signal confirmation, gated email verification, first-passkey WebAuthn registration, passkey authentication sessions, bounded account recovery, and session-authenticated passkey management. Controlled confirmation, email verification, setup-grant passkey registration, and recovery grants are not public authentication or login sessions.",
    "title": "TOWN API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/account/passkeys": {
      "get": {
        "description": "Returns active passkeys for the session account. Exposes only opaque public passkey ids and safe metadata. Requires an active web or mobile session. SetupGrant, RecoveryGrant, and Bearer are rejected.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "passkeys": {
                          "items": {
                            "additionalProperties": false,
                            "properties": {
                              "backedUp": {
                                "anyOf": [
                                  {
                                    "type": "boolean"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "backupEligible": {
                                "anyOf": [
                                  {
                                    "type": "boolean"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "createdAt": {
                                "minLength": 1,
                                "type": "string"
                              },
                              "currentSessionCredential": {
                                "type": "boolean"
                              },
                              "deviceType": {
                                "anyOf": [
                                  {
                                    "enum": [
                                      "multiDevice"
                                    ],
                                    "type": "string"
                                  },
                                  {
                                    "enum": [
                                      "singleDevice"
                                    ],
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "id": {
                                "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
                                "type": "string"
                              },
                              "label": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "lastUsedAt": {
                                "anyOf": [
                                  {
                                    "minLength": 1,
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "label",
                              "createdAt",
                              "lastUsedAt",
                              "deviceType",
                              "backupEligible",
                              "backedUp",
                              "currentSessionCredential"
                            ],
                            "type": "object"
                          },
                          "type": "array"
                        }
                      },
                      "required": [
                        "passkeys"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "statusCode": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "sessionAuth": []
          },
          {
            "mobileSessionAuth": []
          }
        ],
        "summary": "List active passkeys for the authenticated account",
        "tags": [
          "Account"
        ]
      }
    },
...
    "/v1/account/security/reauthentication/passkeys/options": {
      "post": {
        "description": "Issues PublicKeyCredentialRequestOptions bound to the current session for manage_passkeys_authenticate. User verification is required. Requires an active session.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {},
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "options": {
                          "additionalProperties": true,
                          "properties": {},
                          "type": "object"
                        },
                        "reauthenticationCeremonyId": {
                          "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$",
                          "type": "string"
                        }
                      },
                      "required": [
                        "reauthenticationCeremonyId",
                        "options"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "statusCode": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "security": [
          {
            "sessionAuth": []
          },
          {
            "mobileSessionAuth": []
          }
        ],
        "summary": "Create passkey reauthentication options for security freshness",
        "tags": [
          "Account"
        ]
      }
    },
...
  },
...
}